Fix: Se elimina la autogeneración del id en la entidad ParkingSpace ya que el usuario debe asignar manualmente el número del ParkingSpace

+++ src/main/java/com/katusoft/parking_a_lot/dto/RegisterResponseDTO.java
@@ -14,7 +14,7 @@ public class RegisterResponseDTO {
     private ParkingType vehicleType;
     private LocalDateTime dateTimeEntrance;
     private LocalDateTime dateTimeDeparture;
-    private Long parkingSpaceNumber;
+    private Integer parkingSpaceNumber;
     private Double totalAmount;
